misc/HtmlTestSuite/results/test-0976.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Canvas descendants focusability</title>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="https://html.spec.whatwg.org/multipage/canvas.html#being-used-as-relevant-canvas-fallback-content" rel="help"/>
<link href="https://html.spec.whatwg.org/multipage/interaction.html#focusable-area" rel="help"/>
<link href="https://github.com/whatwg/html/issues/7534" rel="help"/>
<meta content="Checks that elements being used as relevant canvas&#xA;    fallback content can't be focusable if they are not in the flat tree." name="assert"/>
</head><body><div id="log"/>
<canvas>
  <section id="shadow-host">
    <button data-focusable="false"/>
    <section data-focusable="false" tabindex="-1">
      <div data-focusable="false" tabindex="-1"/>
      <span data-focusable="false" tabindex="-1"/>
      <a data-focusable="false" href="#"/>
    </section>
  </section>
</canvas>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
setup(() =&gt; {
  const canvas = document.querySelector("canvas");
  assert_greater_than(canvas.getClientRects().length, 0, "Canvas is rendered");
  const shadowHost = document.getElementById("shadow-host");
  const shadowRoot = shadowHost.attachShadow({ mode: "open" });
  const slot = document.createElement("slot");
  slot.name = "slot";
  shadowRoot.appendChild(slot);
});
for (let element of document.querySelectorAll("[data-focusable]")) {
  let title = element.cloneNode(false).outerHTML.toLowerCase();
  title = title.slice(0, title.lastIndexOf("&lt;"));
  test(function() {
    assert_equals(element.getClientRects().length, 0, "Not rendered");
    assert_true(document.activeElement !== element, "Not initially focused");
    element.focus();
    if (JSON.parse(element.dataset.focusable)) {
      assert_true(document.activeElement === element, "Should be focused");
    } else {
      assert_true(document.activeElement !== element, "Shouldn't be focused");
    }
  }, title);
}
</script>
</body></html>
